agent/requirements.txt
playwright
crawl4ai
serpapi
pillow
nest-asyncio
google-search-results
selenium
langchain>=0.2.0
langchain-aws
langchain-openai
boto3
python-dotenv
openai
langgraph
faiss-cpu
pandas
langchain-community
tldextract
furl
